work on timer, implementing clock drift model
git-svn-id: file:///home/svn/symphony/trunk@268 881c09cc-3da8-4082-bb0d-d8ce5b87ff19

+++ ns-3.14/src/symphony/model/tos-node.h
@@ -101,7 +101,7 @@ namespace ns3
     uint32_t
     getNow();
     uint32_t
-    wrapFire(uint32_t);
+    wrapFire(uint64_t);
     /**
      * \returns the unique id of this node.
      *
@@ -114,7 +114,7 @@ namespace ns3
     /**
      * Callback function
      */
-    Callback<uint32_t, uint32_t> callBackFromClock;
+    Callback<uint32_t, uint64_t> callBackFromClock;
 
     Ptr<SimuClock>
     getClock();
